Special:ImportTranslations: Avoid early processing of new translation
When importing translations, the resulting message always
reported "There were no changes for this language." because the methods
to import translations from file were executed twice. This resulted in
the changes not being recognised the second time roundand hence the
aforementioned message is displayed. This change solves this issue and
essentially puts the 'review' stage of the import to actual work.

Bug: T230000
Change-Id: I60b524ed73ef1d984a71925874c8a8886a35d34b