PageTranslationSpecialPage: More compact output of unit ID errors
Instead of outputting one message box per error, store the errors in a
Status object, and output a single message box if there are any errors.
Also add a documentation comment for the method so that others (or
future me) will understand it more easily. Since there are no uses of
this method outside of this class, make it private.

Bug: T306257
Change-Id: I58b532d624a73a90c23b22b035ba011d996f133d