TTMServerAid::getData: Do not swallow TranslationHelperException
ElasticSearchTTMServer::query method throws a
TranslationHelperException that may contain important information if
systems are not working properly.

This purpose of this patch is to surface any exception related to
ElasticSearchTTMServer in order to debug the linked ticket.
Swallowing all exceptions is not a good practise, so we should still
look at a more elegant long term fix.

Bug: T306233
Change-Id: Ifee7da87311357607ab36ec11c7ab14a478d1002