Avoid recreation of translation pages on bulk deletion of pages
We have a function trigger on the hook ArticleDeleteComplete that
tracks translation unit pages deleted, and then triggers a
translation update job.

Update the code to keep track of translatable bundles and translation
pages being deleted and if a translation unit is under these then avoid
triggering a translation page update.

Also pass the trigger action to TranslateRenderJob and avoid recreating
the translation page, if the trigger action for rendering is delete.

Bug: T291724
Change-Id: Ibc59b2f2d704cad36c4ad47ef5e2dce890d21da4