Strip tvar from source when preprocessing
Follow-up from I654a13c0b27fb2951b6cf04d72858d6e39c246b4. That patch
only handled template arguments that had <translate> in it - which is
correct in the context of a full parse, but not necessarily in the
context of template arguments (for which the <translate> tags may be
in the external scope).
This patch also strips the <tvar> tags.

Bug: T296392
Change-Id: Icf5aca8d92872a5fdcf4fe32bb273ce0753f0188