Fix handling of falsy content in Special:ManageMessageGroups
There was a bug in in TranslateUtils::getContentForTitle which
incorrectly converted falsy values like "" or "0" into null.

Since nulls are still possible for other reasons, convert them
to strings explicitly when creating Content.

Better handling these cases is left as future improvement.

This fixes: PHP Notice:  TextContent constructed with $text =
NULL! This may indicate an error in the caller's scope.

This also fixes default value for such messages from "fuzzy" to
"import" (in the source language), since it was doing strict
comparison and null is not equal to "" or "0".

Change-Id: I10c4d8a1b0f6cd54ae5c12a520ce84f916dd3280