MessageTable: Prevent overflow with Vector 2022 CSS grid layout
This was caused when the VectorGrid feature was enabled on wikis:
Iab7e29b8549cfbc97d007e0130853cba26638dbf

The CSS grid layout does not appear to be imposing a max-width or
overflow wrap of any sort. The spans (display:inline) inside
tux-list-message do not honor the width of their parent. When the
message content is long, these expand and make the entire
mw-content-container wide, also increasing the width of the entire
grid layout.

This patch fixes the issue with Special:Translate and
Special:TranslationStash

Bug: T313671
Change-Id: Ia4ff931259be156e582fd3196daac945cc7f6f7a