Remove custom job de-duplication logic on translation unit deletion
The previous approach of tracking the TranslateRenderJob when
translation units was not working because we were not removing jobs
from the static variable that was tracking them.

Use the inbuilt de-duplication logic available in the job queue

Additionally, update the TranslateRenderJob to log the translation
unit that triggered it on save and deletion.

Bug: T303701
Change-Id: I97d9de4773719dcd53106871d0f35ad1165fbdac