TranslateRcFilter: Avoid duplicate columns being queried
Combined with I1320d0373996 it is broken in production and giving this
error:
Error 1060: Duplicate column name 'rc_title'
Here is the query: P32459

If you look at the query, you'd see rc_title and rc_namespace have been
repeated. It's not really needed as RC itself adds those fields but in
order to be sure, let's just use array_merge to make it safe.

Tested on mwdebug1001 and it shows this fixes the issue.

Change-Id: I9c1ca248d12af2f23c57565c97ce08065fe71a78