Add warning when a variant of source language is selected
Sometimes users translate into their language while having a variant of
the source language selected. This is especially common for
British English.

This patch adds a warning message on Special:Translate to prompt
translators that the language currently selected on the target language
selector is a variant of the message group's source language

Bug: T317134
Change-Id: Ic9343d88569b2d44535b75efeda226acf239662a