Add @covers tags to all tests
With PHPUnit 9, tests without @covers tags are considered risky and emit
warnings. Not having @covers is bad practice anyway, so use
@coversNothing instead to make the intention clear, and re-enable the
phpcs rule.

Note: I tried to guess what the test is trying to cover, and this may or
may not be correct in some cases. I guess it doesn't really matter, as
this is already an improvement, and wrong covers shouldn't be worse than
no covers.

Bug: T243600
Change-Id: I5544b83aa3ca105bd5462256a2d5d853425f37a4