Qualify conditions in RC filter for change tags
This ensures that you can both select a change tag to show
and select no rendered page translations at the same time without
getting an SQL error.

The primary fix for this issue is I56a349028471c. When testing,
it seemed like disambiguating the join conditions was not strictly
necessary, but nonetheless it seems like a good idea to do so
anyways.

Bug: T325872
Change-Id: If8d6d8e2dbb4ce166f95e716b48d9ef25c8e13e4