AggregateGroups: Allow setting a source language
The source language of an aggregate group is now selectable as an
optional field in Special:AggregateGroups. This requires that associated
message groups are of the same source language as well; if not, a
error message is shown and the user will not be able to save an
aggregate group with message groups with different source languages.

Bug: T281073
Change-Id: I089804693a857bded22a4ba6253e8b2a3097a19e