MessageGroupStats: Fix attributes for the entity selector label
The empty for attribute causes a warning on Firefox that an
Empty string passed to getElementById(). Update it to point to the
correct item.

The click event handler for the label causes improper behavior
when searching for entities as it tries to put focus on an element
that is not visible.

Bug: T300483
Change-Id: Ie67ae8cd0cf9d796342698903eed6305e461517e