ImportTranslations: Do not add FUZZY if message already has fuzzy tag
When comparing the message being imported with existing message, if the
imported message is fuzzy, then the fuzzy string is prefixed to it. But
the existing message in MessageCollection does not have the fuzzy
string, but is identified as fuzzy with a tag. Fix this by normalizing
strings before comparison.

Bug: T323430
Change-Id: I475b6cfa675a6956e463d428e308854d5f12867c