Use TranslatablePage source language code when fetching fallback
When a template is transcluded we fetch the RevisionRecord for the
translation if available, else we fetch the source language
template. The source language is determined using the MessageGroup.
When a page is marked for translation, it takes some time for the
MessageGroup cache to update. This ends up causing a null
RevisionRecord to be returned during transclusion.

Instead of using the MessageGroup, use source language from the
TranslatablePage since it uses Title::getPageLanguage()

Bug: T323863
Change-Id: Ifff68f4ac37bb7f12eb15f72dcb7bd89c65b1d19