Special:ExportTranslations: EntitySelector to select group
In the event that the Group dropdown on Special:ExportTranalations has a
a lot of message groups, the loading of the dropdown takes a lot of
time. This fix uses the EntitySelector to speed up the selection by
allowing the user to enter a search term that is then debounced.
The results of the search terms are then populated to allow a user to
select a group.

Bug: T331843
Change-Id: I822b3a7b0d33128bbc79f100413e482ecd538815