Check if page has translate markup before updating tp:tag revision
If the page does not have the <translate> markup, we should not be
updating the rt_revision in rev_tag table to match the latest page
revision_id.

Another approach that was considered was comparing the stored revision
id against the penultimate revision id but this approach is a bit
flimsy. For example with WikiImporter, hooks don't get triggered so we
would then have to check (< latest revision record - 1)

Bug: T333068
Change-Id: I1d6322235ed69312d631c46ba216f18a18208f8d