QueryMessageCollectionActionApi: Check for empty filter
At times, the a filter item in the query parameters may be an empty
string. This would cause PHP Notice: Uninitialized string offset: 0 when
an attempt to act on the string is made, such as $filter[0]. To fix
this, we have to ignore empty strings that come in as part of an array
of filters.

Bug: T335517
Change-Id: I0df340c64a670dce051e4ac073fd1c603b0723d4