Use translated language name in tux placeholder
Using the autonym in this message causes a weird mix of languages
(e.g. "Your translation into 日本語") when the interface language
is not the same as the target language. This can lead to bad visual
rendering if the interface language and the target language do not
share the same directionality.

By using mw.language.names, we can use the translated language name
instead of the autonym if it exists.

Change-Id: Iacfe2edc39dcc31b73ad1fac422b617b261bc2fe