Ensure translation unit title is a valid MediaWiki title
When marking a page for translation, validate that the translation
unit page that will be created would result in a valid MediaWiki
title. If not, do not allow the page to be marked for translation
and inform the user so that they can rename the unit or the
translatable page title itself.

Bug: T319375
Change-Id: I1dabc3e6376f20e52658ee83c9c9a7b0adfbea57