Add leave redirect option when moving a translatable page
When moving a translatable page, we provide a checkbox option that adds
redirects for everything except translation unit pages and translation
pages (translatable page, talk pages, regular subpages and their talk
pages).

The checkbox is added to leave a redirect on Special:MovePage when the
page is being moved.

Also add functionality to leave redirect to
MoveTranslatableBundleMaintenanceScript.php

Bug: T62920
Change-Id: I5c1c4b19129b4d4c935529b6431fc7292f1225fd