Improve page translation progress indicators
TLDR;
* Use CSS inspired conic-gradient to 'draw' the progress indicators.
* Remove unneeded resource images.

As it is, the progress indicators on the <languages> list are rendered
using .png images. This patch introduces using of conical-gradient,
a property of CSS, to 'draw' the progress indicators. The chosen color
here is blue. Different percentages of the translation per language
take different forms; with one extreme being a simple hollow blue
circle with 1px border, and the other being a filled blue circle - to
represent 0% and 100% respectively.

Bug: T329046
Change-Id: If4973ec8a130bb049a3138168f4702943c995578