*SpecialPage::getDescription: Return a Message, not a string
The behavior of returning a Message is changing in 40f1694fb268
(MW 1.41). After this change the message will be escaped, but on older
MediaWiki versions it will not be. However, as none of these Messages
were escaped beforehand, this behaviour is acceptable.

We are waiting on this patch to hard-deprecate returning a string
in MediaWiki (which is not escaped).

Bug: T343994
Needed-By: I1059c1da61e751ff958e73987deda45b0368642b
Change-Id: Ibb9bb64ad4440d0eddd1a5a7e66855bc66aeaba8