Add SubsetMessageGroup
Based on MediaWikiTopMessagesGroup in translatewiki.net repo.

Motivation is to avoid PHP code in that repo. We can have it as
a generic feature in Translate, where it gets more attention.

While at it, solve a couple of technical debt tasks by adding caches to
improve performance and recursion guards to prevent infinite recursion.

Bug: T89804
Bug: T208430
Change-Id: I78eb0fc015a3618f255295783355c7047c1250ed