Do not show Mint suggestions on text with wikitext
When translating text with signifant amount of wikitext markup in
Translate, MinT suggestions in produces unrefined results. e.g.
the wikitext may not have closing brace or one in the wrong position.

This patch skips over text with wikitext markup but allows showing of
suggestions of text without wikitext in the translation aids panel on
the TUX interface.

Bug: T349375
Change-Id: I2ebe36c9c72ce0924996f5e060e1e5fbad9d8ccf