Unmark "Allow translation of page title" by default for templates
When a template is marked for translation, make the "Allow translation
of page title" box be unmarked by default. Templates are not usually
meant to be viewed directly by the end user, so translating the
template title is usually a waste of translators' time (since the
translation won't be seen), and could potentially also be
confusing for those seeking to use that template (if the display
title doesn't match the actual template title).

Bug: T305240
Change-Id: I2b02ad74436285cbab6430cdd8b7de92ef5b7534