Use translate_cache.tc_exptime as mw timestamp
The abstract schema defines the field as mwtimestamp,
but it is used as integer with a unix timestamp (ref 0ddac9f2).

Use of wfTimestamp makes this backward compatible between old rows and
new rows as wfTimestamp would detect old rows still as unix and converts
unix to unix, while the new rows would be converted from mw-timestamp to
unix on read for the PersistentCacheEntry.
There are no compares with the timestamp in sql.

This improves postgres support.

Change-Id: I39bb676dfbb26457927b9395a7ed39e99fed6ba3