Ensure translation section markers maintain spaces in their name
In the event that a translation unit contains a section marker such as
<!--hello world--> its ID is always passed as 'hello_world'. This
obviously doesn't match 'hello world' which prevents marking a
translation that should not be invalidated.

This patch fixes the issue by removing the underscore in the ID and
replacing it with a space. As the Translate extension is designed, the
section markers should not contain underscores to begin with. Thus, we
don't have to be concerned about a user using an underscore in their
section marker(s).

Bug: T350222
Change-Id: I4d6e06062823510e0eab90d58fc10f2128593562