Ensure a page context is set for wfMessage
Expanding the `tpt-languages-nonzero` message on mw:Extension:AdvancedSearch/tr
is prompting a "duplicate-args-warning" tracking category from PPFrame_Hash,
presumably because the message in some language is misusing {{PLURAL}},
although it's a little hard to tell for sure.  In any case, we're getting
logspam about it because the wfMessage() invocation doesn't set a page
title context, which the tracking category code expects.

Bug: T354202
Change-Id: I8736c82da29818fe582405fa2042e7fd28e851a2