Consider language selector an auxiliary page search element
The language selector is being included in the general searchable
content of the page, which also makes it available to be snippeted
as part of the search result display. This isn't particularly useful,
mark the selector as `navigation-not-searchable` which excludes it
from search entirely. This seems like the most appropriate action as
the element is used for navigation purposes and searches for the
individual language names should probably not find pages with
translations to that language.

Bug: T352915
Change-Id: I1dcbb770883f960cb1115154b86838d53e51d9fd