ImportTranslatableBundle: Mark page for translation automatically
Use TranslatablePageMarker to mark the page for translation
automatically.

Introduce new parameters to allow importer to specify metadata such
as priority languages.

Eventually we want to be able to read this metadata directly from
the exported file, and these options can still be used to override
the values from the exported file.

Bug: T347155
Change-Id: I82a5be64277c630a10d33a38ca78da574efd42c5