Introduce TranslatableBundleImportTitleFactory
Use that as the title factory when the import is in progress if the
target name option is provided. Use PageTitleRenamer to rename the
titles.

The target name option can be used to specify the namespace and
root-page option, thus removing the need to have those parameters
specified separately.

For now this class does not support importing pages from source wiki
that have namespace that does not exist in the target wiki.

Bug: T347153
Bug: T347154
Change-Id: Id69304b24a158c66b7c5bc7e45d79c37d4309141