MessageIndex: improve rebuilding
Only fetch groups after getting a lock, to avoid stale data.
I believe originally it was outside to keep the critical section
as short as possible, but fetching groups should be pretty fast.

Also move interim cache update inside the critical section to ensure
it is (more) atomic. This should be fast as well.

Bug: T320220
Change-Id: I530d5f317521424c3cc14df26b6204cffe9823c0