GroupSyncCache: Mark group for review even if import strategy is none
For MediaWiki import strategy is true, where all changes need to be
manually reviewed. In this case if there were changes waiting for
review, the group was not marked as "in review" in the group sync
cache.

Update the code to fix this, and move logic to handle import strategy
into ExternalMessageSourceStateImporter.

Bug: T359518
Change-Id: I11d2da3eb03daee484578e45059f61916d602dc2