Tux: Improve handling of save errors
Change the API call when saving a translation in the Translate
user interface, so that errors from AbuseFilter and other places
are returned parsed and in the user's language.

As part of this, add support for multiple simultaneous errors and
displaying them all to the users at the same time.

Also improve readability in some error messages (fix missing definite
articles).

Bug: T360447
Co-Authored-by: Tacsipacsi <tacsipacsi@jnet.hu>
Change-Id: Ic02615486c44058a80963429ac8ebf62c7763b34