Add missing parameter to 'translate-checks-value-not-present'
The message used $1 for both the list of permitted values
and the PLURAL parameter. Since $1 is not a number, this
leads to the PLURAL function always returning the default
number (plural in the case of English) instead of returning
the value for 1 if there is one item in the set.

The $2 parameter is already set inside MatchSetValidator.php,
line 55, so only the message in en.json (and qqq.json) needs
a change to fix this.

Change-Id: Ie318693d9bcd0b408e2bdf8babfbab700439beb8