QueryMessageCollectionApi: Always perform disabled language check
Perform the disabled language check even if the source language
check fails.

Also fix a small pre-existing JavaScript bug I discovered while
testing this. The bug was difficult to reproduce before, since there
was no sure-fire way to trigger an error in the form, but now it
might be easy to trigger an error (just switch to a disabled language
and switch back to an un-disabled language), so the bug would become
more noticeable.

Bug: T364215
Change-Id: I0c914b6b69c288c9350cb0ed4dffcf767b1e17c2