Fix error when marking a new page for translations
Issue was caused by Database::getCacheSetOptions which caused
set to to fail with error: [memcached] Rejected set() for
wiki:translate-mg:translatable-pages due to pending writes.

Added guard that we do not add setOpts when explicitly recaching.
This fixes the error as stopgap solution, but we probably should
review this code with a caching expert given we have had so many
challenges with it.

Bug: T364522
Change-Id: I8aa3e9deba265eaee44dee49f1a061723f19fdb7