Entity selector: Remove apiRequestHook
Since I5fef8cb30874fdc2c935960a429837b54b3436a0, it's become easier to
filter for message groups using the entity selector. We no longer need
the custom code to do so for Special:Special:AggregateGroups.

An unintended side-effect from this change is that
'allowSuggestionsWhenEmpty' would now cause an error as the 'query'
param was not allowed as a an empty string. The solution here is to set
the default param to 'query' = '' in TranslationEntitySearchActionApi.

This patch also limits the number of results returned to 50.

Change-Id: I896bf4682ac5281ebf4309fdb7445dcd8cb9a75a