Simplify message group caching
Explicit regeneration is not supported by WANObjectCache and us
trying to do that anyway seems to be the root of our caching
bugs about translatable pages.

This is an attempt to change the code to follow normal cache
invalidation pattern for hot keys (lockTSE + checkKey).

Do note that there are two levels of caching:
* Each message group type caches minimalized serialized data
* In-process cache of all instantiated message groups tied to the
  MessageGroup singleton

I've changed the code in TranslatablePageMarker to not rely on the
particular group being present in the cache.

I also changed recache to load the latest uncached data for any
other post-processing during the request that may rely on the list
of groups.

Added waitForReplication to RebuildMessageIndexJob like we
already have for many other jobs.

Both message index and message group statistics are self-healing
in the sense that any other update later on will correct
previously stale data. For message group stats though the any
other update may not happen ever for any particular group.

Bug: T320220
Change-Id: I857aa63426fa2c0af01334fc44674190bb5eac18