TranslatableBundleMover: Use DB transaction when moving page
While moving pages for T366979, we encountered an SQL error due to
target pages being modified manually. This resulted in the move
process ending abruptly. The move was completed using moveBatch.php
instead.

Looking at the code for moveBatch.php, it appears to be using a
DB transaction to ensure that the process of moving the page is
atomic. Use the same approach here.

Change-Id: I02e71482cd5ac76ce5cc088a47c167dff7a22233