TranslatablePage: Store source page ids as string in WAN cache
Storing the message bundle and translatable page ids as comma
separated strings in the cache as this results in slightly
smaller size. Also adding a comma at the beginning and end in order
to make contains search simpler.

On commons, this reduces the cache size from 35,653 to 21,548
serialized characters.

Increase pcTTL duration by multiple of 2 for long running processes.

Bug: T366455
Change-Id: I1d5af304bdc1d3313c555e14a187aa722cd81723