TranslateEditAddons: Use user-provided summary in null revision
If the user provided an edit summary in the null edit that unfuzzies a
translation, use that instead of the generic message (“Marked
translation unit as no longer outdated with no changes”), assuming that
it provides explanation on why the translation didn’t need any updates.
Continue using the generic edit message if no custom edit summary was
provided.

Bug: T371502
Change-Id: I338d6171df6b39c95a052d3d2043137c9bae8851