Make ext.translate.quickedit night mode compatible
Use design tokens to make ext.translate.quickedit (the legacy
translation interface) night mode compatible. Remove the link color
setting, as the default link color should suffice in night mode, and is
much better in dark mode. The colors slightly changed, but the
background and border color changes are almost unnoticeable (the link
color change is noticeable, but hopefully it’s for the better).
Set the previously unset text color to make sure that no incompatible
text color is inherited (before this patch, the main problem was that in
night mode, a light text color was inherited, but a light background
color was set).

Change-Id: I868f8549fa4fcda00984b1d518ceb51d8da269a7