SubsetMessageGroup: use structured logging
In production, this hopefully directs the messages to a place where
they’re more likely to be noticed (note that this class is not used in
WMF production, but it *is* used on translatewiki.net). During testing,
it lets us confirm that the right things are logged (for which a
TODO-like comment already existed) and at the same time avoid spamming
the test output.

Bug: T372575
Change-Id: Iad59400edaf2752df9c456dee5ee3711d0a43865