Use OutputPage::addCategoryLinks() instead of ::setCategoryLinks()
The Translate extension is the only user of this deprecated OutputPage
method, and the replacement ::addCategoryLinks() has been present for
at least a decade (d606322c83887c22091820c213a26e78b8a3bfc2 added
type hints for it).  We don't need to clear previously existing categories
in OutputPage because the ParserAfterTidy hook already calls
`$parserOutput->setCategories([])` to clear them in the ParserOutput.

Change-Id: Id25041a7891f588ffa787fdd2c092342eecd30c8