TranslatablePageView: Use recentchanges to check for recent edits
The recentchanges table has lesser number of records compared to
the revision table.

Also avoid the additional join to the actor table by using the
rc_actor column in recentchanges table. Fetching the actor id for
the user may result in an additional query, but there is a high
probability that this is already cached.

Bug: T370265
Change-Id: I8ad3e640589e755526ea2db78984dd6255ae6c8e