Send notification to subscribers of aggregate groups
If a group under an aggregate groups is changed, then subscribers of
that aggregate group should recieve notifications.

At the same time, direct subscribers of the message group who might
also be subscribed to the aggregate group should not receive the
notification again.

Here's another example that describes the complexity,

Imagine an aggregate group AG1 that has message groups MG1, MG2, MG3
There is a change made in MG1, and MG2. Notifications will be sent to
direct subscribers of MG1 and MG2. AG1 subscribers who are not
subscribed to MG1 and MG2 will also be sent a notification.

Bug: T375376
Change-Id: I3e1a58d24c1be215e060dce2121d09e9a631c239