Avoid duplicate `wgLogTypes` entries
The `HookHandler` adds log types to the global `$wgLogTypes` without
first checking to see if they are already present. If the handler
is run more than once (for example in `TranslateSandboxTest`),
subsequent renders of `Special:Log` will generate error messages
in the error log because of the duplicate types.

Filter the `$wgLogTypes` array at the end of the HookHandler to
remove duplicates.

Bug: T375483
Change-Id: Ie9c18b002860e5766037a32cce8ec8331389b5d7