Shorten dataset used in MessageIndexTest
testMessageIndexImplementation uses a dataset of 49286 messages but then
only acts upon 10 of them.

When running tests in parallel over  8 PHPUnit processes, with multiple
extensions and with a MySQL datadir limited to 320MB, we found out the
insertion to use all the thin InnoDB table space.

This change truncate the dataset to 128 messages which has been randomly
picked up using:
```
<?php
$messages = unserialize(
	file_get_contents('tests/phpunit/data/messageindexdata.ser')
);
$newMessages = [];
foreach (array_rand($messages, 128) as $draw) {
	if (is_string($messages[$draw])) {
		$newMessages[$draw] = str_rot13($messages[$draw]);
		continue;
	}
	$newMessages[$draw] = [];
	foreach ($messages[$draw] as $key => $value) {
		$newMessages[$draw][$key] = str_rot13($value);
	}
}
file_put_contents(
	'tests/phpunit/data/messageindexdata.ser',
	serialize($newMessages)
);
```

The MessageIndexTest previously took 2 seconds on my machine and it is
now down to 200 miliseconds.

Bug: T375756
Change-Id: I029944a24071c39c2e4bf4f2797174dddb1e5840