Remove `array_unique` application to global variable
In Ie9c18b002860e5766037a32cce8ec8331389b5d7 we introduced a call
to `array_unique` on `$wgLogTypes` to prevent duplicate log types
appearing in the array. This is too broad - a single extension
should not be rewriting the whole `$wgLogTypes` array as that can
have unknown consequences for other extensions and make bugs
harder to track down.

Remove the `array_unique` call and instead check to see if values
are already present in the array before adding them.

Bug: T375483
Change-Id: I355c18680f0e491a59cea53a64475e033fe676ad