MessageGroupSubscription: Send notifications to aggregate groups
Aggregate message groups can be nested. The approach that we'd taken
earlier in I3e1a58d24c1be215e060dce2121d09e9a631c239 did not account
for this.

Update the code to handle arrays of groups recursively.

We avoid using MessageIndex to determine the groups for a message
because the message index may not be upto date when the notifications
are being sent out.

Bug: T375376
Change-Id: Idf55b3d6e0643d02f8bf8f281b384e1b802a786b