TranslationsSpecialPage: Stop extending SpecialAllPages
SpecialAllPages is not @stable to extend. We don’t really use it either:
a single one-line method is for what we relied on it. Copy over that
method and extend its parent class IncludableSpecialPage.

Use $this->getConfig() instead of a global variable, as that’s better
practice anyway, but it’s now also enforced by PHP_CodeSniffer due to
the new parent class.

Bug: T265309
Change-Id: I9ebfc51a521fb3cbf912a697b33528cf65ecfa3a