Make the group selector readable in night mode
The hotfix in 84f6ebaf9716ca812ba8e5acc5133f170f332045 didn’t even work
around the group selector, as it’s attached to the end of the body, well
outside of the element that got a `skin-invert` class in that patch. Use
Less variables to fix the display. This causes minor changes in the day
mode colors:

- Header (the background behind the tabs and the search field): almost
  unnoticeable change `#f0f0f0` → `#eaecf0`.
- Hover color of groups: almost unnoticeable change `#f8f8f8` →
  `#f8f9fa`.
- Borders between the groups: noticeable but small change `#eee` →
  `#dadde3`. I chose the lightest background color available, but it
  still got a bit darker.
- “N subgroups” text for aggregate groups in the groups list: perfect
  match `#72777d` → `#72777d`.

Also set the text color explicitly to avoid problems arising from colors
being set on `<body>` for whatever reason.

Bug: T371071
Change-Id: Ibaa98bb1bcc4f109de47e5cb90a0ffe959c581f5