Make Special:ManageGroups work on wikis other than Translatewiki
The JavaScript code responsible for handling message renames assumed
that the last piece of the path part of the URL is same as the page name
if and only if no subpage was provided (i.e. we’re processing the
default import). This assumption could fail at (at least) two points:

- The page name and the URL part are equal only if the page name
  contains only ASCII characters. For example, in my Hungarian-language
  wiki, the special page name is `Speciális:ManageMessageGroups`, but
  the URL part is `Speci%C3%A1lis:ManageMessageGroups`.
- The page name may not appear at all in the URL if short URLs [1] are
  not used.

To fix both, pass the import name in a `data-` attribute to the
frontend. This attribute is present even if no subpage was provided,
simplifying the processing of its value.

As far as I see, the title wasn’t used for any other purpose, so skip it
from the form.

[1] https://www.mediawiki.org/wiki/Manual:Short_URL

Change-Id: I60aa1605dcaa23e1885bd84a182b28e8642f6ca9