Use `use function` in Utilities\Yaml
By default, PHP searches for functions in the current namespace. If it
doesn’t find a function in the current namespace, it continues searching
in the global space. This is usually fine (although a bit less efficient
due to the extra lookup); however, if it doesn’t find the function in
the global space either, the error message mentions the local namespace,
which is confusing if the function was meant to be imported from the
global space.

In Utilities\Yaml, missing functions are to be expected: ideally, at
least one of the two supported YAML libraries are installed, but it may
happen that neither of them are. In that case, the wiki administrator is
faced with that error message. Explicitly use the functions from the
global space to make those error messages easier to understand (the
function name MediaWiki\Extension\Translate\Utilities\spyc_load, which
appears without that, can make one think it’s a bug in the extension,
even though it’s actually because of wrong configuration/setup).

Change-Id: I989d329b44f4f4ee8a51da2fa16a5735df99e5e3