Special:TranslationStats: Use HTML5 inputs
The single OOUI input in the otherwise native HTML form didn’t look
nice, and the number inputs with `<input type="text">` were difficult to
use and easy to get wrong (generic keyboard on touchscreen devices, no
client-side validation). On top of this, the descriptions says “All
values have upper and lower limits”, but the form didn’t tell the user
what the upper and lower limits of these number inputs were (it simply
clapped the values at the limits on server side).

To make the form better-looking and easier to use, convert it to a pure
HTML5 form: the date input is `<input type="date">`, the number inputs
are `<input type="number">`, the radio buttons remain radio buttons.
(The language and group selectors are not pure HTML5 in implementation,
but they are pure HTML5 in design.)

This conversion makes the client-side handling a lot easier: no need to
set up and keep track of the OOUI date selector, no need to parse its
input, no need to convert strings to numbers; all these are done by the
browser for us. This means about 70 lines less in the JS file.

The date handling is a bit tricky, though: the API delivering the stats
expects a timestamp (an ISO 8601 date/time string, interpreted in UTC),
but the UI allowed only selecting a date (but in a bit buggy way: if I
selected 23 October from the date selector, and submitted the form, the
next page load showed that I selected 22 October, presumably because it
mixed up UTC and CEST, which is my time zone right now). Keep the UI
variant on the special page, i.e. allow selecting a date, because if we
allow the user to select a date only, it’s less confusing that we
interpret that date in UTC rather than in the user’s time zone. This
means we need to parse the date in the special page backend on our own,
especially because we need to support the old format as well (due to
already-bookmarked URLs and more importantly due to already-embedded
special pages), but this keeps the client-side form clean.

The OOUI date selector used to have a max date of today, but the backend
didn’t use to validate it. While the best practice is that client-side
validation shouldn’t be relied upon, keep this behavior by adding a
`max` attribute but otherwise not caring about the value on the backend:
the best practice is about security-related checks, and this doesn’t
seem like a security-sensitive bit – if the user works around the form
and sets a future date, they’ll simply get all zeros.

Change-Id: I15ecb0efaf92e4072c7d20bda80fb74e3e834e1c