Fix only one page translation language link appearing
This got accidentally broken in core change I8a8d499a6a6646bc86.
Instead of reverting that, better to fix it here.

Mainly switching from "x-pagetranslation" code of the placeholders
to $code-x-pagetranslation to avoid collapsing them all to one.

Small cleanups to the code for readability and removing one unused
key (percentage) in the data.

Bug: T377131
Change-Id: I0e23f7a80350d4641719866748d0d7b48d9440a8