Fix update when a virtual domain is enabled
Use addExtensionUpdateOnVirtualDomain() exclusively when updating tables
on the virtual-translate domain. Split the patch file
translate_cache-alter-varbinary.sql into a translate_cache patch (which
is on the virtual domain), and a patch for everything else.

This domain points to x1 in production, so to run
LoadExtensionSchemaUpdates in production, as is required by the linked
task, it is necessary to install the tables into the correct cluster.

Bug: T352113
Change-Id: Ibc7e9dd0cd2adcecfd76173c1a48dcba143ebfd1